deps: add glm files and update premake config

+++ Projects/Core/Include/Util/PlatformUtils.h
@@ -16,7 +16,7 @@ namespace Ephemeral
     class Time
     {
     public:
-        static float GetTime()
+        static double GetTime()
         {
             return glfwGetTime();
         };
